Moved the Main class into a Logic folder fixed some comments
